+++ CHANGELOG.md
@@ -44,6 +44,7 @@ v5.10.0
 - Fix 3rdparty compilation with ninja/unix makefiles/macos/win (!2537) 
 - Fix Boost, FFTW, VTK, and Zlib thirdparty compilation on macos, fix serial PETsc thirdparty compilation, fix OCE depedencies (!2538)
 - Add scaling factor to limit the Newton step (!2533)
+- Acctivate the inexact forcing in the Newton process (!2534)
 
 **IncNavierStokesSolver**
 - Added an option to process Reynolds Stress fields at a higher polynomial order (!2303)